Minor fixes

+++ mvvm_hybrid/src/main/java/com/mayada1994/moviewatchlist_hybrid/fragments/WatchlistFragment.kt
@@ -26,6 +26,8 @@ class WatchlistFragment : Fragment() {
 
     private val viewModel by viewModels<WatchlistViewModel> { WatchlistComponent.viewModelFactory }
 
+    private var deleteDialog: AlertDialog? = null
+
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
@@ -93,20 +95,24 @@ class WatchlistFragment : Fragment() {
     }
 
     private fun showDeleteMoviesDialog() {
-        val dialogView = DialogEditWatchlistBinding.inflate(layoutInflater)
-        val alertDialog = AlertDialog.Builder(requireContext()).setView(dialogView.root).create()
-
-        with(dialogView) {
-            txtPrompt.text = getString(R.string.dialog_delete_movie_prompt_message)
-            btnOk.setOnClickListener {
-                alertDialog.dismiss()
-                viewModel.deleteMovies()
+        if (deleteDialog == null) {
+            val dialogView = DialogEditWatchlistBinding.inflate(layoutInflater)
+            val alertDialog =
+                AlertDialog.Builder(requireContext()).setView(dialogView.root).create()
+            deleteDialog = alertDialog
+            with(dialogView) {
+                txtPrompt.text = getString(R.string.dialog_delete_movie_prompt_message)
+                btnOk.setOnClickListener {
+                    alertDialog.dismiss()
+                    viewModel.deleteMovies()
+                }
+                btnCancel.setOnClickListener {
+                    alertDialog.dismiss()
+                }
             }
-            btnCancel.setOnClickListener {
-                alertDialog.dismiss()
-            }
+            alertDialog.setOnDismissListener { deleteDialog = null }
+            alertDialog.show()
         }
-        alertDialog.show()
     }
 
     private fun goToSearchScreen() {
